Move non-git functions outside git.go (to repo.go)
git.go should only contain purely git-related functions. Higher level
functions that abstract multiple git and annex commands are found in
repo.go.

The whole file/package and function layout needs some redesign so this
is subject to change.
