+++ repo/repo.go
@@ -4,8 +4,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"net"
 	"net/http"
 	"os"
+	"strconv"
+	"strings"
+	"time"
+
+	"golang.org/x/crypto/ssh"
+	"golang.org/x/crypto/ssh/agent"
 
 	"github.com/G-Node/gin-cli/util"
 	"github.com/G-Node/gin-cli/web"
@@ -26,6 +33,113 @@ type Client struct {
 // NewClient returns a new client for the repo server.
 func NewClient(host string) *Client {
 	return &Client{Client: web.NewClient(host)}
+}
+
+// Temporary (SSH key) file handling
+var privKeyFile util.TempFile
+
+// MakeTempKeyPair creates a temporary key pair and stores it in a temporary directory.
+// It also sets the global tempFile for use by the annex commands. The key pair is returned directly.
+func (repocl *Client) MakeTempKeyPair() (*util.KeyPair, error) {
+	tempKeyPair, err := util.MakeKeyPair()
+	if err != nil {
+		return nil, err
+	}
+
+	description := fmt.Sprintf("tmpkey@%s", strconv.FormatInt(time.Now().Unix(), 10))
+	pubkey := fmt.Sprintf("%s %s", strings.TrimSpace(tempKeyPair.Public), description)
+	authcl := auth.NewClient(repocl.KeyHost)
+	err = authcl.AddKey(pubkey, description, true)
+	if err != nil {
+		return tempKeyPair, err
+	}
+
+	privKeyFile, err = util.SaveTempKeyFile(tempKeyPair.Private)
+	if err != nil {
+		return tempKeyPair, err
+	}
+
+	privKeyFile.Active = true
+
+	return tempKeyPair, nil
+}
+
+// CleanUpTemp deletes the temporary directory which holds the temporary private key if it exists.
+func CleanUpTemp() {
+	if privKeyFile.Active {
+		privKeyFile.Delete()
+	}
+}
+
+func publicKeyFile(file string) ssh.AuthMethod {
+	buffer, err := ioutil.ReadFile(file)
+	if err != nil {
+		return nil
+	}
+
+	key, err := ssh.ParsePrivateKey(buffer)
+	if err != nil {
+		return nil
+	}
+	return ssh.PublicKeys(key)
+}
+
+// Connect opens a connection to the git server. This is used to validate credentials
+// and generate temporary keys on demand, without performing a git operation.
+// On Unix systems, the function will attempt to use the system's SSH agent.
+// If no agent is running or the keys offered by the agent are not valid for the server,
+// a temporary key pair is generated, the public key is uploaded to the auth server,
+// and the private key is stored internally, to be used for subsequent functions.
+func (repocl *Client) Connect() error {
+	util.LogWrite("Checking connection to git server")
+	sshAgent, err := net.Dial("unix", os.Getenv("SSH_AUTH_SOCK"))
+	if err != nil {
+		// No agent running - use temp keys
+		util.LogWrite("No agent running. Setting up temporary keys")
+		_, err = repocl.MakeTempKeyPair()
+		if err != nil {
+			return fmt.Errorf("Error while creating temporary key for connection: %s", err.Error())
+		}
+		return nil
+	}
+
+	agent := ssh.PublicKeysCallback(agent.NewClient(sshAgent).Signers)
+
+	sshConfig := &ssh.ClientConfig{
+		User: repocl.GitUser,
+		Auth: []ssh.AuthMethod{
+			agent,
+		},
+		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
+	}
+
+	util.LogWrite("Attempting connection with key from SSH agent")
+	connection, err := ssh.Dial("tcp", repocl.GitHost, sshConfig)
+	if err != nil && strings.Contains(err.Error(), "unable to authenticate") {
+		// Agent key authentication failed - use temp keys
+		util.LogWrite("Auth key authentication failed. Setting up temporary keys")
+		_, err = repocl.MakeTempKeyPair()
+		if err != nil {
+			return fmt.Errorf("Error while creating temporary key for connection: %s", err.Error())
+		}
+		return nil
+	}
+	// TODO: Attempt connection again after temp key is set up
+
+	if err != nil {
+		// Connection error (other than "unable to auth")
+		return fmt.Errorf("Failed to connect to git host: %s", err.Error())
+	}
+	defer connection.Close()
+
+	session, err := connection.NewSession()
+	util.LogWrite("Creating SSH session")
+	if err != nil {
+		return fmt.Errorf("Failed to create session: %s", err.Error())
+	}
+	defer session.Close()
+	util.LogWrite("Connection to git server OK")
+	return nil
 }
 
 // GetRepo retrieves the information of a repository.
@@ -262,4 +376,237 @@ func (repocl *Client) CloneRepo(repoPath string) (string, error) {
 	}
 
 	return repoName, nil
+}
+
+// FileStatus represents the state a file is in with respect to local and remote changes.
+type FileStatus uint8
+
+const (
+	// Synced indicates that an annexed file is synced between local and remote
+	Synced FileStatus = iota
+	// NoContent indicates that a file represents an annexed file that has not had its contents synced yet
+	NoContent
+	// Modified indicatres that a file has local modifications that have not been committed
+	Modified
+	// LocalChanges indicates that a file has local, committed modifications that have not been pushed
+	LocalChanges
+	// RemoteChanges indicates that a file has remote modifications that have not been pulled
+	RemoteChanges
+	// Unlocked indicates that a file is being tracked and is unlocked for editing
+	Unlocked
+	// Untracked indicates that a file is not being tracked by neither git nor git annex
+	Untracked
+)
+
+// Description returns the long description of the file status
+func (fs FileStatus) Description() string {
+	switch {
+	case fs == Synced:
+		return "Synced"
+	case fs == NoContent:
+		return "No local content"
+	case fs == Modified:
+		return "Locally modified (unsaved)"
+	case fs == LocalChanges:
+		return "Locally modified (not uploaded)"
+	case fs == RemoteChanges:
+		return "Remotely modified (not downloaded)"
+	case fs == Unlocked:
+		return "Unlocked for editing"
+	case fs == Untracked:
+		return "Untracked"
+	default:
+		return "Unknown"
+	}
+}
+
+// Abbrev returns the two-letter abbrevation of the file status
+// OK (Synced), NC (NoContent), MD (Modified), LC (LocalUpdates), RC (RemoteUpdates), UL (Unlocked), ?? (Untracked)
+func (fs FileStatus) Abbrev() string {
+	switch {
+	case fs == Synced:
+		return "OK"
+	case fs == NoContent:
+		return "NC"
+	case fs == Modified:
+		return "MD"
+	case fs == LocalChanges:
+		return "LC"
+	case fs == RemoteChanges:
+		return "RC"
+	case fs == Unlocked:
+		return "UL"
+	case fs == Untracked:
+		return "??"
+	default:
+		return "??"
+	}
+}
+
+// FileStatusSlice is a slice of FileStatus which implements Len() and Less() to allow sorting.
+type FileStatusSlice []FileStatus
+
+// Len is the number of elements in FileStatusSlice.
+func (fsSlice FileStatusSlice) Len() int {
+	return len(fsSlice)
+}
+
+// Swap swaps the elements with intexes i and j.
+func (fsSlice FileStatusSlice) Swap(i, j int) {
+	fsSlice[i], fsSlice[j] = fsSlice[j], fsSlice[i]
+}
+
+// Less reports whether the element with index i should sort before the element with index j.
+func (fsSlice FileStatusSlice) Less(i, j int) bool {
+	return fsSlice[i] < fsSlice[j]
+}
+
+func lfDirect(paths ...string) (map[string]FileStatus, error) {
+	statuses := make(map[string]FileStatus)
+
+	wiResults, _ := AnnexWhereis(paths)
+	for _, status := range wiResults {
+		fname := status.File
+		for _, remote := range status.Whereis {
+			// if no remotes are "here", the file is NoContent
+			statuses[fname] = NoContent
+			if remote.Here {
+				if len(status.Whereis) > 1 {
+					statuses[fname] = Synced
+				} else {
+					statuses[fname] = LocalChanges
+				}
+				break
+			}
+		}
+	}
+
+	asargs := paths
+	if len(asargs) == 0 {
+		// AnnexStatus with no arguments defaults to root directory, so we should use "." instead
+		asargs = []string{"."}
+	}
+	annexstatuses, _ := AnnexStatus(asargs...)
+	for _, stat := range annexstatuses {
+		if stat.Status == "?" {
+			statuses[stat.File] = Untracked
+		} else if stat.Status == "M" {
+			statuses[stat.File] = Modified
+		}
+	}
+
+	return statuses, nil
+}
+
+func lfIndirect(paths ...string) (map[string]FileStatus, error) {
+	statuses := make(map[string]FileStatus)
+
+	gitlsfiles := func(option string) []string {
+		gitargs := []string{"ls-files", option}
+		gitargs = append(gitargs, paths...)
+		stdout, stderr, err := RunGitCommand(gitargs...)
+		if err != nil {
+			util.LogWrite("Error during git ls-files %s", option)
+			util.LogWrite("[stdout]\r\n%s", stdout.String())
+			util.LogWrite("[stderr]\r\n%s", stderr.String())
+			// ignoring error and continuing
+		}
+		var fnames []string
+		for _, fname := range strings.Split(stdout.String(), "\n") {
+			// filter out emtpty lines
+			if len(fname) > 0 {
+				fnames = append(fnames, fname)
+			}
+		}
+		return fnames
+	}
+
+	// Collect checked in files
+	cachedfiles := gitlsfiles("--cached")
+
+	// Collect modified files
+	modifiedfiles := gitlsfiles("--modified")
+
+	// Collect untracked files
+	untrackedfiles := gitlsfiles("--others")
+
+	// Run whereis on cached files
+	wiResults, err := AnnexWhereis(cachedfiles)
+	if err == nil {
+		for _, status := range wiResults {
+			fname := status.File
+			for _, remote := range status.Whereis {
+				// if no remotes are "here", the file is NoContent
+				statuses[fname] = NoContent
+				if remote.Here {
+					if len(status.Whereis) > 1 {
+						statuses[fname] = Synced
+					} else {
+						statuses[fname] = LocalChanges
+					}
+					break
+				}
+			}
+		}
+	}
+
+	// If cached files are diff from upstream, mark as LocalChanges
+	diffargs := []string{"diff", "--name-only", "--relative", "@{upstream}"}
+	diffargs = append(diffargs, cachedfiles...)
+	stdout, stderr, err := RunGitCommand(diffargs...)
+	if err != nil {
+		util.LogWrite("Error during diff command for status")
+		util.LogWrite("[stdout]\r\n%s", stdout.String())
+		util.LogWrite("[stderr]\r\n%s", stderr.String())
+		// ignoring error and continuing
+	}
+
+	diffresults := strings.Split(stdout.String(), "\n")
+	for _, fname := range diffresults {
+		// Two notes:
+		//		1. There will definitely be overlap here with the same status in annex (not a problem)
+		//		2. The diff might be due to remote or local changes, but for now we're going to assume local
+		if strings.TrimSpace(fname) != "" {
+			statuses[fname] = LocalChanges
+		}
+	}
+
+	// Add leftover cached files to the map
+	for _, fname := range cachedfiles {
+		if _, ok := statuses[fname]; !ok {
+			statuses[fname] = Synced
+		}
+	}
+
+	// Add modified and untracked files to the map
+	for _, fname := range modifiedfiles {
+		statuses[fname] = Modified
+	}
+
+	// Check if modified files are actually annex unlocked instead
+	mdfilestatus, err := AnnexStatus(modifiedfiles...)
+	if err != nil {
+		util.LogWrite("Error during annex status while searching for unlocked files")
+	}
+	for _, stat := range mdfilestatus {
+		if stat.Status == "T" {
+			statuses[stat.File] = Unlocked
+		}
+	}
+
+	// Add untracked files to the map
+	for _, fname := range untrackedfiles {
+		statuses[fname] = Untracked
+	}
+
+	return statuses, nil
+}
+
+// ListFiles lists the files and directories specified by paths and their sync status.
+// Setting the Workingdir package global affects the working directory in which the command is executed.
+func ListFiles(paths ...string) (map[string]FileStatus, error) {
+	if IsDirect() {
+		return lfDirect(paths...)
+	}
+	return lfIndirect(paths...)
 }
